step-01/read_data_service/pyproject.toml
[tool.poetry]
name = "read-data-service"
version = "0.1.0"
description = ""
authors = ["usmanashrf <usmanashraf1510@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.18"
pyjwt = "^2.8.0"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
